Move ThemeControl into traits; use #[autoimpl(for<T: trait + ?Sized> Box<T>)]
